change homepagecontroller, allow cors for it and some refactor

+++ hostmanager-backend/src/main/java/ru/serovmp/hostmanager/repository/HostRepository.java
@@ -17,5 +17,5 @@ public interface HostRepository extends JpaRepository<Host, Long> {
     List<Host> recent(@Param("recentLimit") int recentLimit);
 
     @Query(value = "SELECT COUNT(*) FROM hosts WHERE NOT is_dir", nativeQuery = true)
-    public int hostsCount();
+    public long hostsCount();
 }
